requirements.txt
streamlit
pandas
numpy
xlsxwriter
python-docx
matplotlib
openpyxl
